Gets vite.js to work again in the multi-module setting
Note that we don't yet have the server started concurrently with the client, but we need the structure to have common code available in both modules

At the moment, package.json is also being duplicated in the client and top-level. This is awkward and will need fixing when I figure out how to get the vite.js plugin for scala.js to look for its package.json in the client subdirectory.